Support firebase  multi-cloud authentication configuration
Token from firebase does not contains clientId (mapped to the
'issued_to' field of the json token) so that authorizers
reading the clientId runs into a NPE when dealing with such token.
We handle it gracefully now, whith those clientId-based
authorizers simply deny access when the clientId is not provided.